Add interface
